Refactor client request logic into a separate class
